pyproject.toml
[tool.poetry]
name = "medicalimageprocessing"
version = "0.1.0"
description = ""
authors = ["alex <dr.p163@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.37.0"
fastapi = "^0.119.0"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
asyncpg = "^0.30.0"
alembic = "^1.17.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
